+++ docs/source/host.rst
@@ -27,6 +27,6 @@ setup to be performed prior to the WSGI application being created.
 For the implementation of the above see :meth:`~gabbi.driver.build_tests`.
 
 .. _wsgi-intercept: https://pypi.python.org/pypi/wsgi_intercept
-.. _intercept examples: https://github.com/cdent/gabbi/blob/master/gabbi/test_intercept.py
-.. _full examples: https://github.com/cdent/gabbi/blob/master/gabbi/gabbits_intercept/google.yaml
-.. _live examples: https://github.com/cdent/gabbi/blob/master/gabbi/test_live.py
+.. _intercept examples: https://github.com/cdent/gabbi/blob/master/gabbi/tests/test_intercept.py
+.. _full examples: https://github.com/cdent/gabbi/blob/master/gabbi/tests/gabbits_intercept/google.yaml
+.. _live examples: https://github.com/cdent/gabbi/blob/master/gabbi/tests/test_live.py
