+++ gabbi/tests/gabbits_intercept/contenttype.yaml
