+++ gabbi/tests/gabbits_intercept/fixture.yaml
