+++ gabbi/tests/gabbits_intercept/json_extensions.yaml
