+++ gabbi/tests/gabbits_intercept/kitten.png
