+++ gabbi/tests/gabbits_intercept/method_shortcut.yaml
