+++ gabbi/tests/gabbits_intercept/self.yaml
